+++ tests/dummy/app/app.js
@@ -1,3 +1,8 @@
+/* eslint-disable ember/no-classic-classes */
+
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
+
 import Application from '@ember/application';
 import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';